% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_prompts.R
\name{add_prompts}
\alias{add_prompts}
\title{Write a Prompt rds File for Turk}
\usage{
add_prompts(projects.path, prompts, ...)
}
\arguments{
\item{prompts}{A vector of promts to show the person that will be judged.}

\item{project.path}{A path the the projects directory.}

\item{path}{Path of where to put the .rds file.}

\item{\ldots}{ignored.}
}
\value{
Returns the \code{project.path} so it can be used with 
\code{make_project} and \code{add_responses} within a \pkg{magrittr} chain.
}
\description{
\code{add_prompts} - Adds the ID enriched prompt .rds file directly to the project directory.
}
\examples{
temp <- tempdir()
prompts <- c("I am sam", "He is greg", "We are us.")
write_prompts(prompts = prompts, path = file.path(temp, "prompts.rds"))
readRDS(file.path(temp, "prompts.rds"))

\dontrun{
## In a magrittr pipeline
library(dplyr)
prompts <- c("I am sam", "He is greg", "We are us.")
responses <- c("Orange", "Yellow", "Green", "Blue", "Red", "Purple")

make_project(
    path = "L:/swiper/shinyTurk/turk_items/ce_question_tagging_validation",
    key.value = 9999
) \%>\%
    add_prompts(prompts) \%>\%
    add_responses(
        responses = responses,
        googleform_url = 'https://docs.google.com/forms/d/1iLiHtE6xFUzYMN8iB3AXD5xRYVY_oI5Cps1uaujZGng', 
        type = "checkbox"        
    )   
}
}

