% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_responses.R
\name{add_responses}
\alias{add_responses}
\title{Write a Response rds File for Turk}
\usage{
add_responses(projects.path, responses, googleform_url, type = "radio", ...)
}
\arguments{
\item{responses}{A vector of responses to eb repeated each time 'next' is clicked.}

\item{googleform_url}{A google form url (google form should be a url to a google form with 
a text box for (a) user, (b) question ID (c) the clicked response in that order).}

\item{type}{The type of response.  Currently (radio is supported).  Plan to add slider 
(numeric) and checkbox types as well.}

\item{project.path}{A path the the projects directory.}

\item{path}{Path of where to put the .rds file.}

\item{\ldots}{Other arguments passed to \code{sliderInput} (\code{c("min", "max", "value", "step")})}
}
\value{
Returns the \code{project.path} so it can be used with 
\code{make_project} and \code{add_prompts} within a \pkg{magrittr} chain.
}
\description{
\code{add_responses} - Adds the list of responses, google form link, and 
response type .rds file directly to the project directory.
}
\examples{
temp <- tempdir()

## radio responses
responses1 <- c("Very Poor", "Poor", "Neutral", "Good", "Very Good")

write_responses(
    responses = responses1, 
    googleform_url = 'https://docs.google.com/forms/d/1iLiHtE6xFUzYMN8iB3AXD5xRYVY_oI5Cps1uaujZGng', 
    path = file.path(temp, "responses_radio.rds"),
    type = "radio"
)
readRDS(file.path(temp, "responses_radio.rds"))

## checkbox responses
responses2 <- c("Orange", "Yellow", "Green", "Blue", "Red", "Purple")

write_responses(
    responses = responses2, 
    googleform_url = 'https://docs.google.com/forms/d/1iLiHtE6xFUzYMN8iB3AXD5xRYVY_oI5Cps1uaujZGng', 
    path = file.path(temp, "responses_checkbox.rds"),
    type = "checkbox"
)
readRDS(file.path(temp, "responses_checkbox.rds"))

## slider responses
write_responses(
    responses = "Rate the statement poor-1 to great-5", 
    googleform_url = 'https://docs.google.com/forms/d/1iLiHtE6xFUzYMN8iB3AXD5xRYVY_oI5Cps1uaujZGng', 
    path = file.path(temp, "responses_slider.rds"),
    type = "slider",
    max = 5, 
    min = 1,
    value = 3,
    step = .001
)
readRDS(file.path(temp, "responses_slider.rds"))

\dontrun{
## In a magrittr pipeline
library(dplyr)
prompts <- c("I am sam", "He is greg", "We are us.")
responses <- c("Orange", "Yellow", "Green", "Blue", "Red", "Purple")

make_project(
    path = "L:/swiper/shinyTurk/turk_items/ce_question_tagging_validation",
    key.value = 9999
) \%>\%
    add_prompts(prompts) \%>\%
    add_responses(
        responses = responses,
        googleform_url = 'https://docs.google.com/forms/d/1iLiHtE6xFUzYMN8iB3AXD5xRYVY_oI5Cps1uaujZGng', 
        type = "checkbox"        
    )   
}
}

