% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{make_project}
\alias{make_project}
\title{Create a Turk Task Project}
\usage{
make_project(path, key.value, key.path = file.path(dirname(path), "key.csv"),
  ...)
}
\arguments{
\item{path}{A path to the projects dirctory plau the project name.}

\item{key.value}{A value (key code) that will be used in to log into the 
project in the shinyTurk.}

\item{key.path}{A path the where `key.csv` is.}

\item{\ldots}{ignored.}
}
\value{
Returns the \code{project.path} so it can be used with 
\code{add_prompts} and \code{add_responses} within a \pkg{magrittr} chain.
}
\description{
Creates a turk task project in the project directory, adds an 
`instructions.md`, and optionally adds a key value to the `key.csv` file.
}
\examples{
\dontrun{
make_project(
    path = "L:/swiper/shinyTurk/turk_items/ce_question_tagging_validation",
    key.value = 9998
)
}

\dontrun{
## In a magrittr pipeline
library(dplyr)
prompts <- c("I am sam", "He is greg", "We are us.")
responses <- c("Orange", "Yellow", "Green", "Blue", "Red", "Purple")

make_project(
    path = "L:/swiper/shinyTurk/turk_items/ce_question_tagging_validation",
    key.value = 9999
) \%>\%
    add_prompts(prompts) \%>\%
    add_responses(
        responses = responses,
        googleform_url = 'https://docs.google.com/forms/d/1iLiHtE6xFUzYMN8iB3AXD5xRYVY_oI5Cps1uaujZGng', 
        type = "checkbox"        
    )   
}
}

