% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_storage.R
\name{make_storage}
\alias{make_storage}
\title{Make a Storage Base for Turk Projects}
\usage{
make_storage(path, ...)
}
\arguments{
\item{path}{A path to the storage location.}

\item{\ldots}{ignored.}
}
\value{
Returns the path to the storage directory invisibly.
}
\description{
Make a storage directory to place related shinyTurk projects.
}
\examples{
\dontrun{
 make_storage('test_storage')
}
}

