% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTurkTools-package.R
\docType{package}
\name{shinyTurkTools}
\alias{package-shinyTurkTools}
\alias{shinyTurkTools}
\alias{shinyTurkTools-package}
\title{Tools to accompany the shinyTurk.}
\description{
Tools to accompany the shinyTurk..
}

