% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncondense_responses.R
\name{uncondense_responses}
\alias{uncondense_responses}
\title{Separate and Concatenate Multiple Responses}
\usage{
uncondense_responses(x, ...)
}
\arguments{
\item{x}{A vector of concatenated responses or a vector of responses to concatenate.}

\item{\ldots}{ignored.}
}
\description{
\code{uncondense_responses} - Separate multiple responses
}
\examples{
(x <- condense_responses(LETTERS[1:5]))
uncondense_responses(x)
}

