% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_prompts.R
\name{write_prompts}
\alias{write_prompts}
\title{Write a Prompt rds File for Turk}
\usage{
write_prompts(prompts, path = "prompts.rds", ...)
}
\description{
\code{write_prompts} - Create an ID enriched prompt data.frame and save as .rds file.
}

