% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_responses.R
\name{write_responses}
\alias{write_responses}
\title{Write a Response rds File for Turk}
\usage{
write_responses(responses, googleform_url, path = "responses.rds",
  type = "radio", ...)
}
\description{
\code{write_responses} - Create a list of responses, google form link, and response type and save as .rds file.
}

